Feature/grpc (#28)
* Create service library for gRPC API

* Add install target to fetch build deps

* Move protobuf stuff into server/service package

* remove old service package

* yes

* Remove service package

* Start gRPC service

* Add make target for evans and add description in README

* Prep for mTLS

* Implement and test GetAvailableMemory

* Change data type for challenge memory to uint64

* Implement and test HaveCapacity for labs

* Remove debug log

* Change memory to int

* service uses API endpoints

* Add file closure and error handling

* Create networkService interface and refactor+move DNS/DHCP services

* Refactor gRPC api and add RemoveLab endpoint

* Add generated gRPC files

* Implement RemoveLab gRPC endpoint

* Creat second test lab

* Create helper function to delete bulk (config) files

* Add dev target

* Refactor to make port a constant

* Correct memory field

+++ daaukins/server/store/store.go
@@ -23,7 +23,7 @@ type ChallengeTemplate struct {
 	Name   string `yaml:"name"`
 	Id     string `yaml:"id"`
 	Image  string `yaml:"image"`
-	Memory int64  `yaml:"memory"`
+	Memory int    `yaml:"memory"`
 }
 
 type Store struct {
